Move GeneratesDocumentationSearchIndexFile into Services namespace
