+++ packages/framework/src/Actions/PostBuildTasks/GenerateSearch.php
@@ -2,7 +2,7 @@
 
 namespace Hyde\Framework\Actions\PostBuildTasks;
 
-use Hyde\Framework\Actions\GeneratesDocumentationSearchIndexFile;
+use Hyde\Framework\Services\GeneratesDocumentationSearchIndexFile;
 use Hyde\Framework\Concerns\InteractsWithDirectories;
 use Hyde\Framework\Contracts\AbstractBuildTask;
 use Hyde\Framework\Hyde;
